BC-8561 - move account as idm-account to domain folder of account module
