+++ apps/server/src/infra/identity-management/keycloak/service/keycloak-identity-management.service.integration.spec.ts
@@ -3,7 +3,7 @@ import { KeycloakModule } from '@infra/identity-management/keycloak/keycloak.mod
 import KeycloakAdminClient from '@keycloak/keycloak-admin-client-cjs/keycloak-admin-client-cjs-index';
 import UserRepresentation from '@keycloak/keycloak-admin-client/lib/defs/userRepresentation';
 import { ObjectId } from '@mikro-orm/mongodb';
-import { IdmAccount, IdmAccountUpdate } from '@modules/account/repo';
+import { IdmAccount, IdmAccountUpdate } from '@modules/account/domain';
 import { ServerModule } from '@modules/server/server.app.module';
 import { HttpModule } from '@nestjs/axios';
 import { Test, TestingModule } from '@nestjs/testing';
