+++ apps/server/src/modules/account/domain/index.ts
@@ -1,4 +1,5 @@
 export * from './account';
 export * from './account-save';
+export * from './idm-account';
 export * from './update-account';
 export * from './update-my-account';
