+++ apps/server/src/modules/account/domain/services/account-idm.service.integration.spec.ts
@@ -6,8 +6,8 @@ import { ObjectId } from '@mikro-orm/mongodb';
 import { ConfigModule } from '@nestjs/config';
 import { Test, TestingModule } from '@nestjs/testing';
 import { v1 } from 'uuid';
-import { AccountSave } from '..';
-import { AccountIdmToDoMapper, AccountIdmToDoMapperDb, IdmAccount } from '../../repo';
+import { AccountSave, IdmAccount } from '..';
+import { AccountIdmToDoMapper, AccountIdmToDoMapperDb } from '../../repo';
 import { AccountServiceIdm } from './account-idm.service';
 import { AbstractAccountService } from './account.service.abstract';
