+++ apps/server/src/modules/account/domain/services/account-idm.service.spec.ts
@@ -6,9 +6,9 @@ import { NotImplementedException } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { Test, TestingModule } from '@nestjs/testing';
 import { EntityNotFoundError } from '@shared/common/error';
-import { Account, AccountSave } from '..';
+import { Account, AccountSave, IdmAccount } from '..';
 import { AccountConfig } from '../../account-config';
-import { AccountIdmToDoMapper, AccountIdmToDoMapperDb, IdmAccount } from '../../repo';
+import { AccountIdmToDoMapper, AccountIdmToDoMapperDb } from '../../repo';
 import { AccountServiceIdm } from './account-idm.service';
 
 describe('AccountIdmService', () => {
