+++ apps/server/src/modules/account/domain/services/account-idm.service.ts
@@ -5,9 +5,9 @@ import { Injectable, NotImplementedException } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config/dist/config.service';
 import { EntityNotFoundError } from '@shared/common/error';
 import { Counted, EntityId } from '@shared/domain/types';
-import { Account, AccountSave } from '..';
+import { Account, AccountSave, IdmAccountUpdate } from '..';
 import { AccountConfig } from '../../account-config';
-import { AccountIdmToDoMapper, IdmAccountUpdate } from '../../repo';
+import { AccountIdmToDoMapper } from '../../repo';
 import { FindAccountByDbcUserIdLoggable, GetOptionalIdmAccountLoggable } from '../error';
 import { AbstractAccountService } from './account.service.abstract';
