+++ apps/server/src/modules/account/domain/services/account.service.integration.spec.ts
@@ -12,8 +12,8 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { cleanupCollections } from '@testing/cleanup-collections';
 import { MongoMemoryDatabaseModule } from '@testing/database';
 import { v1 } from 'uuid';
-import { Account, AccountSave } from '..';
-import { AccountEntity, AccountIdmToDoMapper, AccountIdmToDoMapperDb, AccountRepo, IdmAccount } from '../../repo';
+import { Account, AccountSave, IdmAccount } from '..';
+import { AccountEntity, AccountIdmToDoMapper, AccountIdmToDoMapperDb, AccountRepo } from '../../repo';
 import { accountFactory } from '../../testing';
 import { AccountServiceDb } from './account-db.service';
 import { AccountServiceIdm } from './account-idm.service';
